+++ requirements.txt
@@ -39,3 +39,5 @@ websockets==10.0
 Werkzeug==2.0.1
 wsproto==1.0.0
 yarl==1.6.3
+
+moviepy~=1.0.3
